[CAMERAWG] - A variety of bug fixes while getting ANDOR camera itself to work in the camerawg code. - Added rep per var and var num controls which were missing - fixed some defaults - added andor initialization code which was missing - fixed a couple of uncaught throws.

+++ Camerawg/PictureSettingsControl.cpp
@@ -212,6 +212,8 @@ void PictureSettingsControl::initialize( cameraPositions& pos, CWnd* parent, int
 	{
 		throw;
 	}
+
+
 }
 
 
@@ -387,7 +389,7 @@ void PictureSettingsControl::handleOptionChange(UINT id, AndorCamera* andorObj)
 		{
 			CString textEdit;
 			exposureEdits[exposureInc].GetWindowTextA(textEdit);
-			int exposure;
+			double exposure;
 			try
 			{
 				exposure = std::stof(std::string(textEdit));
